[MM-103]: Added the test cases for the Zoom meeting in threads and Create new meeting option. (#35)
Summary
This PR consists the test cases for the following scenarios,

Creating a Zoom meeting in the thread using slash command.
Creating a Zoom meeting in the thread using Zoom icon in the app bar.
Clicking on the Create new meeting option when there is already a meeting created and not ended.